Refactor database integration to use Diesel ORM
- Updated Cargo.toml to remove sqlx dependencies and add diesel with r2d2 support.
- Modified database connection management to use Diesel's connection pooling.
- Refactored database migration scripts to use Diesel's query builder.
- Updated user and room mapping models to use Diesel's ORM features.
- Changed database store implementations to utilize Diesel for CRUD operations.
- Adjusted web handlers to accommodate new database interaction methods.
- Added schema definitions for Diesel to manage database tables.
- Cleaned up code formatting and improved error handling in various modules.

+++ migrations/001_initial.sql
@@ -1,6 +1,6 @@
 -- 用户映射表
 CREATE TABLE IF NOT EXISTS user_mappings (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     matrix_user_id TEXT NOT NULL UNIQUE,
     discord_user_id TEXT NOT NULL UNIQUE,
     discord_username TEXT NOT NULL,
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS user_mappings (
 
 -- 房间映射表
 CREATE TABLE IF NOT EXISTS room_mappings (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     matrix_room_id TEXT NOT NULL UNIQUE,
     discord_channel_id TEXT NOT NULL UNIQUE,
     discord_channel_name TEXT NOT NULL,
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS room_mappings (
 
 -- 事件跟踪表
 CREATE TABLE IF NOT EXISTS processed_events (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     event_id TEXT NOT NULL UNIQUE,
     event_type TEXT NOT NULL,
     source TEXT NOT NULL, -- 'matrix' or 'discord'
@@ -32,8 +32,8 @@ CREATE TABLE IF NOT EXISTS processed_events (
 
 -- 用户活动表
 CREATE TABLE IF NOT EXISTS user_activity (
-    id SERIAL PRIMARY KEY,
-    user_mapping_id INTEGER NOT NULL REFERENCES user_mappings(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_mapping_id BIGINT NOT NULL REFERENCES user_mappings(id) ON DELETE CASCADE,
     activity_type TEXT NOT NULL,
     timestamp TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
     metadata JSONB
